fix production build
