+++ apps/genomicsdb/app/genome-browser/page.tsx
@@ -1,4 +1,4 @@
-import { UnderConstructionPage } from "@/components/UnderConstructionPage";
+import { UnderConstructionPage } from "@/components/pages/UnderConstructionPage/UnderConstructionPage";
 
 export default function GenomeBrowser() {
     return <UnderConstructionPage source={"Genome Browser"} />;
