+++ apps/genomicsdb/app/tutorials/page.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { UnderConstructionPage } from "@/components/UnderConstructionPage";
+import { UnderConstructionPage } from "@/components/pages/UnderConstructionPage/UnderConstructionPage";
+
 // import { TutorialsPage } from "@/components/pages/tutorials-page";
 
 export default function TutorialsPage() {
